// src/core/http/generated/models/payment.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { Invoice } from './invoice.ts';
import type { PaymentMetadata } from './paymentMetadata.ts';
import type { PaymentStatus } from './paymentStatus.ts';
import type { Subscription } from './subscription.ts';
import type { Tenant } from './tenant.ts';

export interface Payment {
  id: string;
  tenant_id?: string;
  tenant?: Tenant;
  subscription_id?: string;
  subscription?: Subscription;
  invoice_id?: string;
  invoice?: Invoice;
  gateway?: string;
  gateway_payment_id?: string;
  gateway_transaction_id?: string;
  status: PaymentStatus;
  method: string;
  amount: number;
  currency: string;
  paid_at?: string;
  failed_at?: string;
  refunded_at?: string;
  failure_reason?: string;
  metadata?: PaymentMetadata;
  created_at: string;
  updated_at?: string;
}